Implement restrictions based on roles and permissions

+++ app/Http/Controllers/CommentController.php
@@ -21,6 +21,9 @@ class CommentController extends Controller
     }
 
     public function destroy(Comment $comment){
+        if ($comment->user_id != Auth::id()){
+            abort(403);
+        }
         $featureId = $comment->feature_id;
         $comment->delete();
         return to_route("feature.show",$featureId) ;
